Change order of setup localstack
The current acceptance pipeline it is too unstable.

Sometimes the test passes and sometimes it fails.

In the setup_test_env.sh we try to create the buckets for the filestore
to use.

The problem is that sometimes the localstack didn't start the whole
stack (s3, Dynamo, etc):

Move the localstack setup (bucket creation) to the last thing in the build.

**But this is not enough because we need to wait for localstack to be
accepting requests**. That's the reason that I added a small loop that
waits for localstack to be running in order to the bucket creation to
be done, which it solves the pipeline and makes the pipeline
*stable and strong*.