Adjust specs for grouped email attachments
[This Submitter PR](https://github.com/ministryofjustice/fb-submitter/pull/317)
changed the behaviour of the submitter to attach as many files to each
email as possible up to the 10MB maximum limit.

These tests now will send one email to each of the service outputs for
the pdf submission and the uploaded file. However at this moment any csv
files will continue to be sent in their own email.