Make the services one specific container
I was having problems to run the forms together with the integration.

First: updating tests and updating the container was taking forever
because it needs to start all forms again.

Moving all forms to your own container makes sense. It increase
the boot time only when you setup the first time.