Test email output against and form in test environment
This changes the email output spec to run against an actual form
(acceptance-tests-email-output) in the dev-test environment.

We want to be able to run acceptance tests against live forms and check
the output sent via AWS SES in order to be sure that the entire platform
is working as it should.

The authentication with the Gmail API is done using the desktop app
method (https://developers.google.com/gmail/api/quickstart/ruby) as
opposed to using the server to server OAuth2 approach. This is mainly
because in order to use the latter a service account is required. This
service account then has to be allowed domain wide delegation. Once it
has this then an organisation administrator needs to grant permissions
for that account to access users emails. The problem with this is that
it potentially allows access to ALL users email inboxes, not just the
acceptance tests specific email address that we use. We consider this
too high risk to do and have therefore continued using the desktop app
approach.

This has a caveat in that it technically is meant to require an actual
user to visit a URL in their browser to obtain a code which is then used
to generate the required access token and refresh token. We manually did
this step and create the token.yaml on the fly with those values. The
expiration of the access token is apparently several years in the
future, however it it does need to be replaced then a user will once
again have to visit the generated URL with the correct scopes in order
to retrieve a code to generate both tokens again. Not ideal, but a
necessary workaround for the moment.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Matt Tei <matt.tei@digital.justice.gov.uk>