Refresh access token during each test suite run
We need to use the refresh token to obtain a new access token in order
to authenticate with the Gmail API.

As a result of doing it this way we no longer need to periodically
regenerate manually, however at some point the refresh token will expire
and another one will have to be regenerated using the developer console.
It is a long lived refresh token so hopefully it will be quite a while.